requirements.txt
openai
streamlit
pillow